instance_segmentation/requirements.txt
ultralytics==8.1.18
pillow
Flask==3.0.2